refactor(auction): 경매 서비스 명칭을 범용화
