+++ src/main/java/io/wisoft/ignoa_api/auction/service/AuctionFacade.java
